add:#47 商品リストを表示するコンポーネント汎用的にして移動
