+++ src/components/uniqueParts/ProductList/ProductCard/hook.ts
