+++ src/components/uniqueParts/ProductList/style.ts
@@ -1,0 +1,9 @@
+import { makeStyles } from '@material-ui/styles';
+
+/** スタイル */
+export const useStyles = makeStyles({
+  grid: {
+    display: 'flex',
+    flexFlow: 'row wrap',
+  },
+});
